Return HttpResult<Listening> from serve()
This can be used to stop the server again, by calling the `close()`
method on the `Listening` instance.

Because Iron doesn't re-publish the `Listening` type, we need to add
Hyper as a dependency. See https://github.com/iron/iron/issues/406.

This change allows us to add integration tests that test the server
output.